refactor: changed the database of mysql to postgresql

+++ src/entity/AmountToReceive.ts
@@ -24,7 +24,7 @@ export default class AmountToReceive {
     @Column('text')
     description: string;
 
-    @Column('double')
+    @Column('float')
     value: number;
 
     @ManyToOne(() => User, (user) => user.amountToReceive, { onDelete: 'CASCADE' })
